Longs

+++ src/util/math/PrimeNumberGenerator.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class PrimeNumberGenerator {
-    public static List<Integer> generatePrimeNumbers(int limit)
+    public static List<Long> generatePrimeNumbers(long limit)
     {
-        List<Integer> primes=new ArrayList<>();
-        for (int i = 2; i < limit+1; i++) {
+        List<Long> primes=new ArrayList<>();
+        for (long i = 2; i < limit+1; i++) {
             boolean wasPrime=true;
-            for(Integer prime:primes)
+            for(Long prime:primes)
             {
                 wasPrime&=i%prime!=0;
                 if(!wasPrime)
